requirements.txt
# Python version: 3.12.12

Pillow==11.3.0
fastapi==0.120.1
numpy==2.0.2
opencv-python==4.12.0.88
pydantic==2.11.10
torch==2.8.0+cu126
transformers==4.57.1
uvicorn==0.38.0
